feat: refactor page components to improve structure and organization, including updates to Home, About, and Playlist pages with new section headers and component imports
